+++ knyst/src/graph.rs
@@ -5,7 +5,6 @@
 //!
 //! ```
 //! # use knyst::prelude::*;
-//! # use knyst::wavetable::*;
 //! let graph_settings = GraphSettings {
 //!     block_size: 64,
 //!     sample_rate: 44100.,
@@ -29,14 +28,10 @@
 use crate::gen::{Gen, GenContext, GenState};
 #[allow(unused)]
 use crate::trig;
-use crate::{BlockSize, Sample, SampleRate};
+use crate::{BlockSize, Sample};
 use knyst_macro::impl_gen;
 // For using impl_gen inside the knyst crate
 use crate as knyst;
-// #[cfg(loom)]
-// use loom::cell::UnsafeCell;
-#[cfg(loom)]
-use loom::sync::atomic::Ordering;
 
 #[macro_use]
 pub mod connection;
@@ -58,7 +53,6 @@ use slotmap::{new_key_type, SecondaryMap, SlotMap};
 use std::cell::UnsafeCell;
 use std::collections::{HashMap, HashSet};
 use std::mem;
-#[cfg(not(loom))]
 use std::sync::atomic::Ordering;
 use std::sync::atomic::{AtomicBool, AtomicU64};
 use std::sync::{Arc, RwLock};
@@ -426,13 +420,14 @@ enum CopyOrAdd {
     Add,
 }
 
-/// One task to complete, for the node graph Safety: Uses raw pointers to nodes
-/// and buffers. A node and its buffers may not be touched from the Graph while
-/// a Task containing pointers to it is running. This is guaranteed by an atomic
-/// generation counter in GraphGen/GraphGenCommunicator to allow the Graph to
-/// free nodes once they are no longer used, and by the Arc pointer to the nodes
-/// owned by both Graph and GraphGen so that if Graph is dropped, the pointers
-/// are still valid.
+/// One task to complete, for the node graph
+///
+/// Safety: Uses raw pointers to nodes and buffers. A node and its buffers may
+/// not be touched from the Graph while a Task containing pointers to it is
+/// running. This is guaranteed by an atomic generation counter in
+/// GraphGen/GraphGenCommunicator to allow the Graph to free nodes once they are
+/// no longer used, and by the Arc pointer to the nodes owned by both Graph and
+/// GraphGen so that if Graph is dropped, the pointers are still valid.
 struct Task {
     /// The node key may be used to send a message to the Graph to free the node in this Task
     node_key: NodeKey,
@@ -1040,7 +1035,6 @@ impl Drop for OwnedRawBuffer {
 /// # Example
 /// ```
 /// use knyst::prelude::*;
-/// use knyst::wavetable::*;
 /// use knyst::graph::RunGraph;
 /// let graph_settings = GraphSettings {
 ///     block_size: 64,
@@ -1070,6 +1064,8 @@ impl Drop for OwnedRawBuffer {
 pub struct Graph {
     id: GraphId,
     name: String,
+    /// The nodes are stored here on the heap, functionally Pinned until dropped.
+    /// The `Arc` guarantees that the `GraphGen` will keep the allocations alive even if the `Graph` is dropped.
     nodes: Arc<UnsafeCell<SlotMap<NodeKey, Node>>>,
     node_keys_to_free_when_safe: Vec<(NodeKey, Arc<AtomicBool>)>,
     buffers_to_free_when_safe: Vec<Arc<OwnedRawBuffer>>,
@@ -3161,25 +3157,39 @@ impl Graph {
             Err(ScheduleError::SchedulerNotCreated)
         }
     }
+    /// Goes through all of the nodes that are connected to nodes in `nodes_to_process` and adds them to the list in
+    /// reverse depth first order.
+    ///
     fn depth_first_search(
         &self,
         visited: &mut HashSet<NodeKey>,
         nodes_to_process: &mut Vec<NodeKey>,
     ) -> Vec<NodeKey> {
-        let mut stack = Vec::with_capacity(self.get_nodes().capacity());
-        while let Some(node_index) = nodes_to_process.pop() {
-            stack.push(node_index);
-            // cloning the input edges here to avoid unsafe
+        let mut node_order = Vec::with_capacity(self.get_nodes().capacity());
+        while !nodes_to_process.is_empty() {
+            let node_index = *nodes_to_process.last().unwrap();
+
             let input_edges = &self.node_input_edges[node_index];
+            let mut found_unvisited = false;
+            // There is probably room for optimisation here by managing to
+            // not iterate the edges multiple times.
             for edge in input_edges {
                 if !visited.contains(&edge.source) {
                     nodes_to_process.push(edge.source);
                     visited.insert(edge.source);
+                    found_unvisited = true;
+                    break;
                 }
             }
+            if !found_unvisited {
+                node_order.push(nodes_to_process.pop().unwrap());
+            }
         }
-        stack
+        node_order
     }
+    /// Looks for the deepest (furthest away from the graph output) node that is also an output node, i.e.
+    /// a node that is both an output node and an input to another node which is eventually connected to
+    /// an output is deeper than a node which is only connected to an output.
     fn get_deepest_output_node(&self, start_node: NodeKey, visited: &HashSet<NodeKey>) -> NodeKey {
         let mut last_connected_node_index = start_node;
         let mut last_connected_output_node_index = start_node;
@@ -3239,7 +3249,7 @@ impl Graph {
         }
 
         let stack = self.depth_first_search(&mut visited, &mut nodes_to_process);
-        self.node_order.extend(stack.into_iter().rev());
+        self.node_order.extend(stack.into_iter());
 
         // Check if feedback nodes need to be added to the node order
         let mut feedback_node_order_addition = vec![];
@@ -3284,7 +3294,7 @@ impl Graph {
             }
         }
         self.node_order
-            .extend(feedback_node_order_addition.into_iter().rev());
+            .extend(feedback_node_order_addition.into_iter());
 
         // Add all remaining nodes. These are not currently connected to anything.
         let mut remaining_nodes = vec![];
@@ -3295,6 +3305,14 @@ impl Graph {
         }
         self.node_order.extend(remaining_nodes.iter());
         self.disconnected_nodes = remaining_nodes;
+        // debug
+        // let nodes = self.get_nodes();
+        // for (i, n) in self.node_order.iter().enumerate() {
+        //     let name = nodes.get(*n).unwrap().name;
+        //     println!("{i}: {name}, {n:?}");
+        // }
+        // dbg!(&self.node_order);
+        // dbg!(&self.disconnected_nodes);
     }
     /// Returns the block size of the [`Graph`], not corrected for oversampling.
     pub fn block_size(&self) -> usize {
@@ -4216,51 +4234,14 @@ impl Mult {
         {
             use std::simd::f32x2;
             let simd_width = 2;
-            for _ in 0..block_size / simd_width {
+            for _ in 0..(*block_size / simd_width) {
                 let s_in0 = f32x2::from_slice(&value0[..simd_width]);
                 let s_in1 = f32x2::from_slice(&value1[..simd_width]);
-                let product = s_in0 * s_in1;
-                product.copy_to_slice(out_buf);
-                in0 = &value0[simd_width..];
-                in1 = &value1[simd_width..];
-                out_buf = &mut out_buf[simd_width..];
-            }
-        }
-        GenState::Continue
-    }
-}
-
-struct NaiveSine {
-    phase: Sample,
-    phase_step: Sample,
-    amp: Sample,
-}
-impl NaiveSine {
-    pub fn update_freq(&mut self, freq: Sample, sample_rate: Sample) {
-        self.phase_step = freq / sample_rate;
-    }
-}
-
-#[impl_gen]
-impl NaiveSine {
-    #[process]
-    fn process(
-        &mut self,
-        freq: &[Sample],
-        amp: &[Sample],
-        sig: &mut [Sample],
-        block_size: BlockSize,
-        sample_rate: SampleRate,
-    ) -> GenState {
-        for i in 0..*block_size {
-            let freq = freq[i];
-            let amp = amp[i];
-            self.update_freq(freq, *sample_rate);
-            self.amp = amp;
-            sig[i] = self.phase.cos() * self.amp;
-            self.phase += self.phase_step;
-            if self.phase > 1.0 {
-                self.phase -= 1.0;
+                let prod = s_in0 * s_in1;
+                prod.copy_to_slice(product);
+                value0 = &value0[simd_width..];
+                value1 = &value1[simd_width..];
+                product = &mut product[simd_width..];
             }
         }
         GenState::Continue